# Save allocations in Zip