--- a/Rx.NET/Source/src/System.Reactive/Linq/Observable/Zip.Generated.tt
+++ b/Rx.NET/Source/src/System.Reactive/Linq/Observable/Zip.Generated.tt
@@ -82,9 +82,8 @@ for (var j = 1; j <= i; j++)
 for (var j = 1; j <= i; j++)
 {
 #>
-                var d<#=j#> = new SingleAssignmentDisposable();
-                disposables[<#=j - 1#>] = d<#=j#>;
-                _observer<#=j#> = new ZipObserver<T<#=j#>>(_gate, this, <#=j - 1#>, d<#=j#>);
+                _observer<#=j#> = new ZipObserver<T<#=j#>>(_gate, this, <#=j - 1#>);
+                disposables[<#=j - 1#>] = _observer<#=j#>;
                 base.Queues[<#= j - 1#>] = _observer<#=j#>.Values;
 
 <#
@@ -93,7 +92,7 @@ for (var j = 1; j <= i; j++)
 for (var j = 1; j <= i; j++)
 {
 #>
-                d<#=j#>.Disposable = source<#=j#>.SubscribeSafe(_observer<#=j#>);
+                _observer<#=j#>.SetResource(source<#=j#>.SubscribeSafe(_observer<#=j#>));
 <#
 }
 #>