A lot of SingleAssignmentDisposables can be saved in Zip.