@danielcweber 2018-06-20
That would be very cool, I'll have a look into this.