@danielcweber 2018-06-20
Would you prefer some local benchmark results (as some kind of proof) in the PR or rather have the benchmarks part of the CI build ?