@danielcweber 2018-06-20
The n-ary tests are really tests for the 3-ary-Zips, the savings are not so impressive (3 allocations per run), but of course that will be much more the greater n is.

```
BenchmarkDotNet=v0.10.14, OS=Windows 10.0.17134
Intel Core i7-6650U CPU 2.20GHz (Skylake), 1 CPU, 4 logical and 2 physical cores
Frequency=2156251 Hz, Resolution=463.7679 ns, Timer=TSC
  [Host]     : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3110.0
  DefaultJob : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3110.0


Original:

                       Method |     Mean |     Error |    StdDev |
----------------------------- |---------:|----------:|----------:|
          Zip_NAry_Asymmetric | 15.02 us | 0.3393 us | 0.5381 us |
 Zip_NAry_Asymmetric_Selector | 14.26 us | 0.3561 us | 0.3497 us |
           Zip_NAry_Symmetric | 12.10 us | 0.1842 us | 0.1633 us |
  Zip_NAry_Symmetric_Selector | 12.80 us | 0.2177 us | 0.2036 us |
   Zip_NAry_Enumerable_Simple | 13.17 us | 0.2549 us | 0.2503 us |




Less allocations:

                       Method |     Mean |     Error |    StdDev |
----------------------------- |---------:|----------:|----------:|
          Zip_NAry_Asymmetric | 14.92 us | 0.2749 us | 0.2437 us |
 Zip_NAry_Asymmetric_Selector | 14.76 us | 0.2923 us | 0.3001 us |
           Zip_NAry_Symmetric | 12.59 us | 0.2504 us | 0.2342 us |
  Zip_NAry_Symmetric_Selector | 13.21 us | 0.2122 us | 0.1772 us |
   Zip_NAry_Enumerable_Simple | 13.42 us | 0.2488 us | 0.2555 us |
```