@danielcweber 2018-06-20
Of course, this is only half of the truth as less allocations means that GC may kick in later and less often.