@danielcweber 2018-06-20
Here we go. Again, we don't see that big a boost here, but another benefit of less allocations is that GC may kick in later and less often, which is not reflected in this benchmark.

```
Original:

                       Method |       Mean |     Error |    StdDev | 
----------------------------- |-----------:|----------:|----------:|
            Zip_AllCompleted2 |   8.352 us | 0.1553 us | 0.1377 us |
            Zip_AllCompleted3 |  11.626 us | 0.1709 us | 0.1515 us |
            Zip_AllCompleted4 |  14.901 us | 0.2862 us | 0.2811 us |
            Zip_AllCompleted5 |  18.526 us | 0.3624 us | 0.3026 us |
            Zip_AllCompleted6 |  24.086 us | 0.4448 us | 0.4161 us |
            Zip_AllCompleted7 |  28.698 us | 0.5023 us | 0.4698 us |
            Zip_AllCompleted8 |  34.896 us | 0.6896 us | 0.6113 us |
            Zip_AllCompleted9 |  41.483 us | 0.7713 us | 0.7575 us |
           Zip_AllCompleted10 |  50.294 us | 0.4446 us | 0.3941 us |
           Zip_AllCompleted11 |  57.110 us | 0.7622 us | 0.7130 us |
           Zip_AllCompleted12 |  66.456 us | 1.3266 us | 1.8159 us |
           Zip_AllCompleted13 |  73.969 us | 1.2653 us | 1.1836 us |
           Zip_AllCompleted14 |  86.634 us | 1.3661 us | 1.2778 us |
           Zip_AllCompleted15 |  96.360 us | 1.8158 us | 1.7834 us |
           Zip_AllCompleted16 | 111.756 us | 2.2157 us | 2.1761 us |


Less allocations:
                       Method |       Mean |     Error |    StdDev |
----------------------------- |-----------:|----------:|----------:|
            Zip_AllCompleted2 |   8.069 us | 0.1520 us | 0.1690 us |
            Zip_AllCompleted3 |  11.132 us | 0.1231 us | 0.1028 us |
            Zip_AllCompleted4 |  14.793 us | 0.2242 us | 0.2097 us |
            Zip_AllCompleted5 |  18.540 us | 0.3074 us | 0.2875 us |
            Zip_AllCompleted6 |  23.367 us | 0.4649 us | 0.4774 us |
            Zip_AllCompleted7 |  29.251 us | 0.3931 us | 0.3677 us |
            Zip_AllCompleted8 |  35.450 us | 0.6922 us | 0.6798 us |
            Zip_AllCompleted9 |  42.327 us | 0.8452 us | 1.0061 us |
           Zip_AllCompleted10 |  49.990 us | 0.9474 us | 1.0137 us |
           Zip_AllCompleted11 |  56.294 us | 0.8105 us | 0.6768 us |
           Zip_AllCompleted12 |  66.008 us | 1.3048 us | 1.1567 us |
           Zip_AllCompleted13 |  77.102 us | 0.8779 us | 0.8212 us |
           Zip_AllCompleted14 |  86.705 us | 1.6075 us | 1.4250 us |
           Zip_AllCompleted15 |  95.842 us | 1.5038 us | 1.4067 us |
           Zip_AllCompleted16 | 109.378 us | 1.7171 us | 1.6062 us |
```