@onovotny 2018-06-20
I'm not sure that benchmarks in a CI build would be really useful given how much it depends on the machine load. I think having it in the PR as a reference is helpful though