@danielcweber 2018-06-20
Agreed, and builds would take forever.

Do you have experience with interpreting the benchmarks? Comparison of the mean is not so impressive but the standard deviation would suggest that the measurements are much less flaky, maybe that's the GC kicking in less frequently?