@danielcweber 2018-06-21
Results with `MemoryDiagnoser`:

```
Original:

                       Method |       Mean |     Error |    StdDev |   Gen 0 | Allocated |
----------------------------- |-----------:|----------:|----------:|--------:|----------:|
            Zip_AllCompleted2 |   7.697 us | 0.0468 us | 0.0437 us | 18.1732 |  37.29 KB |
            Zip_AllCompleted3 |  10.558 us | 0.0739 us | 0.0655 us | 19.5007 |  40.08 KB |
            Zip_AllCompleted4 |  13.607 us | 0.0438 us | 0.0388 us | 20.8282 |  42.76 KB |
            Zip_AllCompleted5 |  16.860 us | 0.1704 us | 0.1594 us | 22.2168 |  45.77 KB |
            Zip_AllCompleted6 |  21.847 us | 0.1523 us | 0.1424 us | 23.8953 |  49.17 KB |
            Zip_AllCompleted7 |  26.215 us | 0.1332 us | 0.1246 us | 25.6348 |  52.89 KB |
            Zip_AllCompleted8 |  32.573 us | 0.4329 us | 0.3615 us | 27.7710 |  56.95 KB |
            Zip_AllCompleted9 |  38.065 us | 0.2881 us | 0.2406 us | 29.8462 |  61.35 KB |
           Zip_AllCompleted10 |  45.841 us | 0.1505 us | 0.1334 us | 32.2266 |  66.13 KB |
           Zip_AllCompleted11 |  51.955 us | 0.3339 us | 0.2960 us | 34.7290 |  71.28 KB |
           Zip_AllCompleted12 |  59.543 us | 0.3348 us | 0.3132 us | 37.3535 |  76.71 KB |
           Zip_AllCompleted13 |  68.253 us | 0.8254 us | 0.7317 us | 39.9170 |  82.49 KB |
           Zip_AllCompleted14 |  79.642 us | 0.6238 us | 0.5835 us | 43.0908 |  88.61 KB |
           Zip_AllCompleted15 |  87.353 us | 0.4121 us | 0.3653 us | 46.2646 |  95.05 KB |
           Zip_AllCompleted16 | 101.971 us | 0.5110 us | 0.4530 us | 49.5605 | 101.81 KB |

Less allocations:

                       Method |       Mean |     Error |    StdDev |   Gen 0 | Allocated |
----------------------------- |-----------:|----------:|----------:|--------:|----------:|
            Zip_AllCompleted2 |   7.513 us | 0.0460 us | 0.0407 us | 18.1732 |  37.29 KB |
            Zip_AllCompleted3 |  10.813 us | 0.0502 us | 0.0445 us | 19.4092 |   39.9 KB |
            Zip_AllCompleted4 |  13.804 us | 0.0682 us | 0.0638 us | 20.7214 |  42.57 KB |
            Zip_AllCompleted5 |  17.298 us | 0.0894 us | 0.0793 us | 22.2168 |  45.56 KB |
            Zip_AllCompleted6 |  22.050 us | 0.1599 us | 0.1495 us | 23.8037 |  48.95 KB |
            Zip_AllCompleted7 |  26.669 us | 0.1567 us | 0.1466 us | 25.6348 |  52.72 KB |
            Zip_AllCompleted8 |  33.237 us | 0.2590 us | 0.2423 us | 27.5269 |  56.66 KB |
            Zip_AllCompleted9 |  39.601 us | 0.4323 us | 0.3832 us | 29.7241 |  61.02 KB |
           Zip_AllCompleted10 |  46.347 us | 0.1650 us | 0.1378 us | 32.0435 |  65.81 KB |
           Zip_AllCompleted11 |  52.834 us | 0.2213 us | 0.1962 us | 34.4238 |   70.9 KB |
           Zip_AllCompleted12 |  60.607 us | 0.2722 us | 0.2273 us | 36.9873 |  76.33 KB |
           Zip_AllCompleted13 |  70.993 us | 0.3774 us | 0.3346 us | 39.9170 |  82.24 KB |
           Zip_AllCompleted14 |  79.750 us | 0.3760 us | 0.3333 us | 42.9688 |   88.2 KB |
           Zip_AllCompleted15 |  90.132 us | 0.5021 us | 0.4696 us | 46.1426 |  94.59 KB |
           Zip_AllCompleted16 | 101.322 us | 0.5323 us | 0.4979 us | 49.3164 | 101.33 KB |
```

Allocation performance has improved, time performance has not.